backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
httpx
pandas
scikit-learn
python-telegram-bot
pytest
